+++ CHANGELOG.md
@@ -1,10 +1,19 @@
+##0.8
+
+Firmware manager API updates, torque settings adding to velocity control and new NSNotifications. Useful Definitions documentation added.
+
+* Firmware manager API has been updated. Apps should now be able to check for updates and switch to the motrr app to apply them.
+* Velocity control now has setting for torque, to prevent stalling at high speeds and improve power consumption at low speeds.
+* Galileo disconnect/connect events can now be observed using NSNotifications as well as the delegate pattern. This permits multiple observers.
+* New documentation added with Useful Definitions. This is a stopgap until appledoc releases support for constants and enums.
+
 ###0.7.6
 
 Revoked current position property for velocity control due to poor accuracy. Feature will be added back at a later time if possible.
 
 ###0.7.5
 
-Added block based interface to position control. Also added missing property currentPosition to velocity control.
+Added block based interface to position control.
 
 ###0.7.4